sitemap, redirections et liens variés 🐐

+++ dodelete.php
@@ -5,6 +5,10 @@
  * Date: 22/11/2017
  * Time: 15:19
  */
+if (!isset($_POST['id'])) {
+    header('Location: index.php?error=nopostdatadelete');
+    exit;
+}
 require_once "connexion.php";
 $requete = "DELETE FROM 
 `garage`
